services/agent-service/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
pymongo==4.6.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
langchain==0.2.6
langchain-community==0.2.6
langchain-ollama==0.1.0
pyjwt==2.8.0
